move parse markdown page logic to separate file, add order and title to md files, move grouping logic from hook into plugin
