+++ docs/core/index.md
@@ -1,0 +1,8 @@
+---
+title: test
+order: 0
+---
+
+# Test index page
+
+testing
