+++ docs/graphcache/Populate.md
@@ -1,3 +1,8 @@
+---
+title: populate
+order: 4
+---
+
 # Populate Exchange
 
 `populate` is an exchange for auto-populating fields in your mutations.
