+++ docs/graphcache/help.md
@@ -1,3 +1,8 @@
+---
+title: help
+order: 2
+---
+
 # Help!
 
 **This document lists out all errors and warnings in `@urql/exchange-graphcache`.**
