+++ docs/graphcache/schema.md
@@ -1,3 +1,8 @@
+---
+title: schema
+order: 6
+---
+
 # Schema
 
 As mentioned in the docs we allow for the schema to be passed
