+++ docs/graphcache/updates.md
@@ -1,3 +1,8 @@
+---
+title: UPDATES
+order: 7
+---
+
 # Updates
 
 When the cache receives a response it will try and do its best to
