+++ packages/site/plugins/source-markdown/hooks.js
@@ -8,23 +8,6 @@ export const useMarkdownPage = () => {
     pages: undefined,
   };
 };
-export const useMarkdownPages = () => {
-  return useRouteData().pages;
-};
 export const useMarkdownTree = () => {
-  return useMarkdownPages().reduce((acc, page) => {
-    const sectionIndex = acc.findIndex(a => a.section === page.section);
-    if (sectionIndex === -1) {
-      return [
-        ...acc,
-        {
-          section: page.section,
-          pages: [page],
-        },
-      ];
-    }
-
-    acc[sectionIndex].pages = [...acc[sectionIndex].pages, page];
-    return acc;
-  }, []);
+  return useRouteData().pages;
 };
