+++ packages/site/plugins/source-markdown/node.api.js
@@ -1,17 +1,12 @@
-import { promisify } from 'util';
 import { createSharedData } from 'react-static/node';
-import { read as readVFile } from 'to-vfile';
 import frontmatter from 'remark-frontmatter';
 import squeeze from 'remark-squeeze-paragraphs';
-import GithubSlugger from 'github-slugger';
-import { select, selectAll } from 'unist-util-select';
 import toString from 'mdast-util-to-string';
 import { parse as yaml } from 'yaml';
 import * as path from 'path';
 import remark from 'remark';
-import nodeGlob from 'glob';
 
-const glob = promisify(nodeGlob);
+import { getPages, groupPages } from './parseMarkdown';
 
 const staticPluginSourceMarkdown = (opts = {}) => ({
   async getRoutes(_, { config }) {
@@ -19,14 +14,14 @@ const staticPluginSourceMarkdown = (opts = {}) => ({
     const location = path.resolve(config.paths.ROOT, opts.location);
 
     // Get page data for each discovered markdown file
-    const markdownPages = await getMarkdownData(
+    const markdownPages = await getPages(
       getMarkdownProcessor(opts.remarkPlugins),
       location,
       opts.pathPrefix
     );
 
     // Share data, since all pages will be displayed e.g. in the sidebar
-    const pages = createSharedData(markdownPages);
+    const groupedPages = createSharedData(groupPages(markdownPages));
 
     // Create react-static routes for each page
     return markdownPages.map(page => ({
@@ -34,7 +29,7 @@ const staticPluginSourceMarkdown = (opts = {}) => ({
       // The markdown file becomes the "template" which the Webpack loader
       // below picks up
       template: `${path.resolve(location, page.originalPath)}.md`,
-      sharedData: { pages },
+      sharedData: { pages: groupedPages },
       getData: () => ({ ...page, headings: undefined }),
     }));
   },
@@ -88,55 +83,6 @@ const getMarkdownProcessor = (plugins = []) => {
   });
 
   return processor;
-};
-
-const getMarkdownData = async (processor, location, pathPrefix = '') => {
-  const slugger = new GithubSlugger();
-
-  // Find all markdown files in the given location
-  const mds = (await glob('**/*.md', { cwd: location })).map(x =>
-    path.resolve(location, x)
-  );
-
-  // Map each markdown to its page data
-  return Promise.all(
-    mds.map(async originalPath => {
-      // Reproduce the current markdown file's route path
-      const relative = path.relative(location, originalPath);
-      const filename = path.basename(relative, '.md');
-      const dirname = path.dirname(relative);
-      const newPath = path.join(pathPrefix, dirname, filename);
-
-      // Parse the given markdown file into MAST
-      const vfile = await readVFile(originalPath);
-      const tree = processor.parse(vfile);
-
-      // Parse the frontmatter yaml data to JSON
-      const frontmatter = yaml(select('yaml', tree)?.value || '') || {};
-
-      // Find all headings and convert them to a reusable format
-      const headings = selectAll('heading', tree).map(node => {
-        const depth = node.depth;
-        const value = (depth === 1 && frontmatter.title) || toString(node);
-        const slug = slugger.slug(value);
-        return { value, slug, depth };
-      });
-
-      // Add fallback for Frontmatter title to first h1 heading
-      frontmatter.title =
-        frontmatter.title ||
-        headings.find(x => x.depth === 1)?.value ||
-        newPath;
-
-      return {
-        section: dirname,
-        originalPath: path.join(dirname, filename),
-        path: newPath,
-        headings,
-        frontmatter,
-      };
-    })
-  );
 };
 
 export default staticPluginSourceMarkdown;
